Add CR** endpoints for transaction use cases;

+++ crm/src/main/java/com/shift/crm/api/models/responses/TransactionResponse.java
@@ -4,5 +4,5 @@ import com.shift.crm.api.models.enums.PaymentType;
 
 import java.time.LocalDateTime;
 
-public record TransactionResponse(Long id, Integer amount, PaymentType paymentType, LocalDateTime transactionDate, SellerShort producer) {
+public record TransactionResponse(Long id, Long amount, PaymentType paymentType, LocalDateTime transactionDate, SellerShort producer) {
 }
